Scenes + Added Character Art
There's still two more scenes, so 5a and 5b are missing at the moment.
That and I need to do scene sprites for everyone beyond the bar scene,
but their portraits for dialogue are there. I made a new camera movement
script but I'm not sure how to make it so that you stop moving at the
edge of the "scene"